Extract CP862 to its own file
